Revert "Feature/component hot area"
