+++ packages/editor/src/components/ComponentItem/index.iframe.scss
@@ -1,5 +1,4 @@
 @import 'styles/helper';
-@import '../HotArea/index';
 
 .vize-component-item {
   position: relative;
@@ -8,9 +7,6 @@
 
   &.selected {
     & > .vize-component-item-mask {
-      border: 2px solid $main-color !important;
-    }
-    .editor-preview-component-hot-areas {
       border: 2px solid $main-color !important;
     }
   }
@@ -37,14 +33,13 @@
   }
 
   .vize-component-item-mask,
-  .editor-preview-component-hot-areas
   .vize-component-item-select-mode-mask {
     position: absolute;
-    top: 0;
-    left: 0;
-    z-index: $min-overlay-z-index;
     width: 100%;
     height: 100%;
+    left: 0;
+    top: 0;
+    z-index: $min-overlay-z-index;
     border: 2px solid $opacity-rgb;
     transition: all ease 200ms;
   }
@@ -57,24 +52,19 @@
       & > span {
         display: inline-block;
       }
-      + .vize-component-event-proxy {
-        .editor-preview-component-hot-area--text{
-          display: inline-block;
-        }
-      }
     }
 
     & > span {
-      position: absolute;
-      bottom: 0;
-      left: 0;
-      display: none;
-      padding: 4px 8px;
       border-top-right-radius: 6px;
+      display: none;
       background-color: $main-color;
       color: white;
+      padding: 4px 8px;
+      position: absolute;
       font-size: 12px;
       line-height: 14px;
+      bottom: 0;
+      left: 0;
       opacity: 0.9;
     }
   }
@@ -92,20 +82,19 @@
         display: flex;
         color: white;
       }
-      
     }
 
     & > span {
+      font-size: 14px;
       display: none;
       align-items: center;
       justify-content: center;
-      font-size: 14px;
       cursor: pointer;
 
       svg {
-        margin-right: 8px;
         font-size: 20px;
         line-height: 20px;
+        margin-right: 8px;
       }
     }
   }
@@ -113,17 +102,15 @@
 
 .vize-container-edit-mode-mask {
   position: fixed;
-  top: 0;
-  left: 0;
-  z-index: $min-overlay-z-index + 1;
   width: 100vw;
   height: 100vh;
+  left: 0;
+  top: 0;
   background-color: rgba(0, 0, 0, 0.4);
+  z-index: $min-overlay-z-index + 1;
 }
 
 // dragging sortable component
 body > .vize-component-item {
   z-index: $max-overlay-z-index;
 }
-
-
