+++ packages/editor/src/components/HotAreaManager/index.tsx
@@ -22,6 +22,8 @@ class State {
 
   loaded = false;
 
+  selectedIndex = -1;
+
   hotAreas: IHotArea[] = [];
 }
 
@@ -69,13 +71,13 @@ export class HotAreaManager extends React.Component {
   private added = false;
 
   public componentDidMount(): void {
-    events.on(EventEmitTypes.MANAGE_HOT_AREA, (instance: ComponentInstance) => {
+    events.on(EventEmitTypes.MANAGE_HOT_AREA, (instance: ComponentInstance, selectedIndex: number) => {
       if (!instance.hotAreas || !instance.data.src) {
         return;
       }
       this.instance = instance;
       this.src = getImageSrc(instance);
-      this.setState({ visible: true, loaded: false });
+      this.setState({ selectedIndex, visible: true, loaded: false });
     });
   }
 
@@ -91,12 +93,12 @@ export class HotAreaManager extends React.Component {
   private onConfirmModal = () => {
     const {
       imgContainerInfo,
-      state: { hotAreas: iHotAreas },
+      state: { hotAreas: iHotAreas, selectedIndex },
     } = this;
     const hotAreas = transformHotAreaFromPxToPercent(iHotAreas, imgContainerInfo);
 
     hotAreasStore.setHotAreas(hotAreas);
-    // selectStore.selectHotArea(selectedIndex);
+    selectStore.selectHotArea(selectedIndex);
     this.onCloseModal();
   };
 
@@ -176,8 +178,8 @@ export class HotAreaManager extends React.Component {
         const hotAreas = [...this.state.hotAreas, movingHotArea];
         this.setState({
           hotAreas,
+          selectedIndex: hotAreas.length - 1,
         });
-        selectStore.selectHotArea(hotAreas.length - 1);
         this.added = true;
       } else {
         this.setSizeAndPosition(this.state.hotAreas.length - 1, {
@@ -192,8 +194,8 @@ export class HotAreaManager extends React.Component {
     const { hotAreas } = this.state;
     this.setState({
       hotAreas: R.remove(index, 1, hotAreas),
+      selectedIndex: -1,
     });
-    selectStore.selectHotArea(-1);
   };
 
   private moveHotArea = (index: number, direction: MoveHotAreaDirection) => {
@@ -226,7 +228,7 @@ export class HotAreaManager extends React.Component {
       moveHotArea,
       setPosition,
       setSizeAndPosition,
-      state: { hotAreas },
+      state: { selectedIndex, hotAreas },
     } = this;
 
     return (
@@ -241,10 +243,8 @@ export class HotAreaManager extends React.Component {
             moveHotArea={moveHotArea}
             setPosition={setPosition}
             setSizeAndPosition={setSizeAndPosition}
-            activated={index === selectStore.hotAreaIndex}
-            onClick={() => {
-              selectStore.selectHotArea(index);
-            }}
+            activated={index === selectedIndex}
+            onClick={() => this.setState({ selectedIndex: index })}
           />
         ))}
       </div>
@@ -256,8 +256,8 @@ export class HotAreaManager extends React.Component {
     const area = iCopyHotArea(hotAreas[index]);
     this.setState({
       hotAreas: [...hotAreas, area],
+      selectedIndex: index,
     });
-    selectStore.selectHotArea(index);
   };
 
   private renderLoading = () => <Spin className="hot-area-manager-loading" />;
