+++ packages/editor/src/states/components.ts
@@ -10,7 +10,6 @@ import {
   componentEventDepsMap,
   ComponentIndex,
   createComponentInstance,
-  deleteComponentNode,
   deleteCurrentPageComponentIndex,
   deletePageComponentInstanceIndexMap,
   DepsType,
@@ -114,7 +113,6 @@ export class ComponentsStore extends StoreWithUtils<ComponentsStore> {
     selectStore.selectPage(selectStore.pageIndex);
     eventStore.deleteDepsEventInstances(DepsType.Component, key);
     componentEventDepsMap.deleteEventDepsMap(key);
-    deleteComponentNode(key);
   };
 
   @action
