+++ packages/editor/src/types/eventListeners.ts
@@ -1,8 +1,6 @@
 /**
  * @desc Component Events
  */
-export const ACTION_TRIGGER_PREFIX = '__impage_action_trigger_';
-
 export enum ComponentEventListenerTypes {
   INIT = 'afterRender',
   CLICK = 'click',
@@ -13,6 +11,3 @@ export enum ComponentEventListenerTypes {
   ENTER_VIEW = 'enterView',
   LEAVE_VIEW = 'leaveView',
 }
-
-export type HotAreaEventListenerTypes = ComponentEventListenerTypes;
-export const HotAreaEventListenerTypes = ComponentEventListenerTypes;
