+++ packages/editor/src/types/index.ts
@@ -5,7 +5,6 @@ export * from './eventListeners';
 export * from './events';
 export * from './global';
 export * from './helper';
-export * from './hotArea';
 export * from './materials';
 export * from './pages';
 export * from './plugins';
