+++ packages/editor/src/types/styles.ts
@@ -96,31 +96,3 @@ export enum FixedOutset {
   BottomLeft = 'BottomLeft',
   BottomRight = 'BottomRight',
 }
-
-export interface MergedCommonStyles {
-  width?: string;
-  height?: string;
-  maxWidth?: string;
-  transform?: string;
-  opacity?: number;
-  borderRadius?: string;
-  color?: string;
-  fontSize?: number;
-  lineHeight?: string;
-  textAlign?: string;
-  fontWeight?: string;
-  margin?: string;
-  padding?: string;
-  border?: string;
-  background?: string;
-  zIndex?: number;
-  backgroundColor?: string;
-  backgroundImage?: string;
-  backgroundSize?: string;
-  backgroundPosition?: string;
-  backgroundRepeat?: string;
-}
-
-export interface MergedStyle extends MergedCommonStyles {
-  [key: string]: any;
-}
