+++ packages/editor/src/utils/common.ts
@@ -92,16 +92,8 @@ export function getMaterialsIdentityName(libName: string, name: string) {
   return `${libName}_${name}`.toLocaleLowerCase();
 }
 
-export function getHotAreaId(key: number): string {
-  return `__vize-hotarea-wrapper-${key}`;
-}
-
 export function px(px: number): string {
   return `${px}px`;
-}
-
-export function percent(percent: number): string {
-  return `${percent}%`;
 }
 
 export function pxWithAuto(px: number | 'auto'): string {
