+++ packages/editor/src/utils/eventEmiter.ts
@@ -8,7 +8,6 @@ export enum EventEmitTypes {
   JUMP_ATTR_EDIT_TAB = 'jump_attr_edit_tab',
   CHANGE_ATTR_EDIT_TAB = 'change_attr_edit_tab',
   COMPONENT_INTERSECTING_CHANGE = 'component_intersecting_change',
-  HOT_AREA_INTERSECTING_CHANGE = 'hot_area_intersecting_change',
   RELOAD_RENDERER = 'reload_renderer',
 }
