--- packages/editor/src/utils/nodeMap/index.ts
@@ -1,2 +1,0 @@
-export * from './componentNodes';
-export * from './hotAreaNodes';
