+++ packages/materials-universal/src/components/Image/index.tsx
@@ -5,10 +5,12 @@ interface Props {
   data: {
     src: string;
   };
-  hotAreas?: React.ReactElement;
+  hotAreas?: React.ReactElement | null;
 }
 
-export default function Image({ data: { src }, hotAreas }: Props) {
+export default function Image({ data, hotAreas }: Props) {
+  const { src } = data;
+
   return (
     <div className="vize-materials-universal image">
       {hotAreas}
